Revert "Quick change to compress and preprocess into double to avoid the cast."
This reverts commit 40e33f3ef8d5833385a2ec566b01e3596e64992d.

+++ src/BayesRRmz.cpp
@@ -167,7 +167,7 @@ int BayesRRmz::runGibbs()
     return 0;
 }
 
-void BayesRRmz::processColumn(unsigned int marker, const Map<VectorXd> &Cx)
+void BayesRRmz::processColumn(unsigned int marker, const Map<VectorXf> &Cx)
 {
     const unsigned int N(data.numKeptInds);
     const double NM1 = double(N - 1);
